requirements.txt
asgiref==3.8.1
build==1.0.3
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.8
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.16.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
more-itertools==10.1.0
msgpack==1.0.7
packaging==24.1
pexpect==4.9.0
pipenv==2024.0.3
pkginfo==1.10.0
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.9.0
pyproject_hooks==1.0.0
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
sqlparse==0.5.1
tomlkit==0.12.1
trove-classifiers==2023.9.19
urllib3==2.0.6
virtualenv==20.26.5
zipp==3.17.0
